# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Snipr URL shortener API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /domains:
    get:
      operationId: getDomains
      tags: [domains]
      summary: Get all custom domains for workspace
      responses:
        "200":
          description: List of domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
    post:
      operationId: createDomain
      tags: [domains]
      summary: Add a custom domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDomainRequest"
      responses:
        "201":
          description: Domain added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "409":
          description: Domain already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /domains/{id}:
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Remove a custom domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /folders:
    get:
      operationId: getFolders
      tags: [folders]
      summary: Get all folders for workspace
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
  /folders/{id}:
    put:
      operationId: updateFolder
      tags: [folders]
      summary: Update a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /conversions:
    post:
      operationId: trackConversion
      tags: [conversions]
      summary: Track a conversion event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackConversionRequest"
      responses:
        "201":
          description: Conversion tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversion"
    get:
      operationId: getConversions
      tags: [conversions]
      summary: List workspace conversions
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Conversions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversionRow"
  /conversions/revenue:
    get:
      operationId: getRevenueReport
      tags: [conversions]
      summary: Revenue and conversion analytics by link and campaign
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Domain:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        domain:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - domain
        - verified
        - createdAt
        - updatedAt
    CreateDomainRequest:
      type: object
      properties:
        domain:
          type: string
      required:
        - domain
...
    Folder:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - color
        - createdAt
        - updatedAt
    CreateFolderRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
      required:
        - name
    UpdateFolderRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
...
    TrackConversionRequest:
      type: object
      properties:
        slug:
          type: string
          nullable: true
        workspaceId:
          type: string
          nullable: true
        linkId:
          type: string
          nullable: true
        clickEventId:
          type: string
          nullable: true
        eventName:
          type: string
        revenue:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        utmCampaign:
          type: string
          nullable: true
        utmSource:
          type: string
          nullable: true
        utmMedium:
          type: string
          nullable: true
    Conversion:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        linkId:
          type: string
          nullable: true
        eventName:
          type: string
        revenue:
          type: string
          nullable: true
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - eventName
        - currency
        - createdAt
    ConversionRow:
      type: object
      properties:
        id:
          type: string
        linkId:
          type: string
          nullable: true
        eventName:
          type: string
        revenue:
          type: string
          nullable: true
        currency:
          type: string
        utmCampaign:
          type: string
          nullable: true
        utmSource:
          type: string
          nullable: true
        utmMedium:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        slug:
          type: string
          nullable: true
        linkTitle:
          type: string
          nullable: true
      required:
        - id
        - eventName
        - currency
        - createdAt
    RevenueByLink:
      type: object
      properties:
        linkId:
          type: string
          nullable: true
        slug:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        conversions:
          type: number
        revenue:
          type: number
      required:
        - conversions
        - revenue
    RevenueByCampaign:
      type: object
      properties:
        campaign:
          type: string
        conversions:
          type: number
        revenue:
          type: number
      required:
        - campaign
        - conversions
        - revenue
    RevenueByEvent:
      type: object
      properties:
        eventName:
          type: string
        conversions:
          type: number
        revenue:
          type: number
      required:
        - eventName
        - conversions
        - revenue
    RevenueReport:
      type: object
      properties:
        totalRevenue:
          type: number
        totalConversions:
          type: number
        byLink:
          type: array
          items:
            $ref: "#/components/schemas/RevenueByLink"
        byCampaign:
          type: array
          items:
            $ref: "#/components/schemas/RevenueByCampaign"
        byEvent:
          type: array
          items:
            $ref: "#/components/schemas/RevenueByEvent"
      required:
        - totalRevenue
        - totalConversions
        - byLink
        - byCampaign
        - byEvent
...