# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Snipr URL shortener API
 * OpenAPI spec version: 0.1.0
 */

export interface Domain {
  id: string;
  workspaceId: string;
  domain: string;
  verified: boolean;
  createdAt: Date;
  updatedAt: Date;
}